+++ R/xrd.get.AllPeaks.R
@@ -1,18 +1,17 @@
-#' Finds All Peaks in a Spectrum
+#' Finds all peaks in an XRD spectrum
 #'
 #' @description
 #' Attempts to find peaks in an XRD spectrum and return a list of angles with
-#' the peak locations. (This function is highly experimental and does not
-#' work well.)
+#' the peak locations. (Use with caution, might find additional peaks.)
 #'
 #'
-#' @param data data frame with intensity and 2 theta angle
-#' @param min.Prominence minimum prominence in percent (optional)
-#' @param Try.Sigma vector with peak widths used to start fitting (optional)
-#' @param deltaTheta search area around main peak
-#' @param Range range to search for peaks (optional)
+#' @param data data frame with intensity and 2 theta angle.
+#' @param min.Prominence minimum prominence in percent (optional).
+#' @param Try.Sigma vector with peak widths used to start fitting (optional).
+#' @param deltaTheta search area around main peak (optional).
+#' @param Range range to search for peaks (optional).
 #'
-#' @return vector with peak positions
+#' @return sorted vector with peak positions with 0.01 degree precision.
 #'
 #' @seealso [xrd.find.Peak()]
 #'
@@ -28,24 +27,23 @@ xrd.get.AllPeaks <- function(data,
                              Try.Sigma = c(0.2,0.1,0.05,0.3),
                              deltaTheta = 5,
                              Range = c(0,90)) {
+  # verify that data has consistent format
   dataXRD <- check_dataXRD(data)
-  # XRD data
+  # use a filter to limit range of data
   d <- xrd_filter(dataXRD, Range[1], Range[2])
 
   step.size = 2 # search peaks with this step size
   theta.min = min(d$TwoTheta) + step.size/2
   theta.max = max(d$TwoTheta) - step.size/2
 
+  pk = c()
+  if ((theta.max-theta.min)<1) {
+    warning("XRD data has insufficient angular range.")
+    return(pk)
+  }
 
   peakPos.list = c(seq(from=theta.min, to=theta.max, by=step.size),
                    seq(from=theta.min+(step.size/2), to=theta.max-(step.size/2), by=step.size))
-
-  pk = c()
-
-  if ((theta.max-theta.min)<1) {
-    warning("Data has insufficient angle range.")
-    return(pk)
-  }
 
   for(ang in peakPos.list) {
     n1 <- xrd_filter(d, ang-deltaTheta, ang+deltaTheta)
@@ -57,5 +55,6 @@ xrd.get.AllPeaks <- function(data,
         pk = c(pk, p1)
     }
   }
-  sort(pk)
+  # sort the peak positions and only keep with a precision of 0.01 deg
+  sort(unique(round(pk,2)))
 }