+++ R/xrd.getSampleFiles.R
@@ -1,10 +1,13 @@
-#' Get XRD File Names with Sample Data
+#' XRD Filenames with Sample Data
 #'
 #' @description
-#' returns sample XRD  from library
+#' Returns filenames with XRD sample data stored in different formats.
+#' These data sets help the user quickly get started with example codes
+#' for this package. It can also be used for testing and debugging purposes.
+#'d
 #'
-#' @param fileExt file extension, could be \code{NA} for all, or "asc", "txt", "ras", "rasx"
-#' @return vector with path/filename to sample XRD data files
+#' @param fileExt XRD filename extension: "asc", "txt", "ras", "rasx"
+#' @return vector with path and filename to sample XRD data files
 #'
 #' @author Thomas Gredig
 #' @examples
@@ -12,10 +15,10 @@
 #' basename(file.list)
 #'
 #' @export
-xrd.getSampleFiles <- function(fileExt = NA) {
+xrd.getSampleFiles <- function(fileExt = NULL) {
   pfad = system.file("extdata",package="rigakuXRD")
   file.list = dir(pfad)
-  if (!is.na(fileExt)) file.list = file.list[grep(paste0(fileExt,"$"), file.list)]
+  if (!is.null(fileExt)) file.list = file.list[grep(paste0(fileExt,"$"), file.list)]
   file.path(pfad, file.list)
 }