+++ R/xrd_check.R
@@ -1,6 +1,6 @@
 #' Check whether we have a xrd S3 class object or vectors
-#' @param TwoTheta either an xrd S3 object or a vector with two theta data
-#' @param Intensity \code{NULL} if xrd S3 object, otherwise intensity vector
+#' @param TwoTheta vector of two theta angles from xrd spectrum OR xrd object.
+#' @param Intensity \code{NULL} if xrd S3 object, otherwise xrd intensity vector
 #' @noRd
 check_dataXRD <- function(TwoTheta, Intensity=NULL) {
   if(inherits(TwoTheta,"xrd")) {
@@ -19,13 +19,13 @@ check_dataXRD <- function(TwoTheta, Intensity=NULL) {
 
 
 
-#' Print information
+#' Verbose output for debugging
 #' @importFrom cli cli_inform
 #' @importFrom glue glue
 #' @noRd
 xrd_inform <- function(msg) {
-  is_debug_mode <- (getOption("verbose", default = FALSE) == TRUE)
-  if (is_debug_mode) {
+  is_verbose_mode <- (getOption("verbose", default = FALSE) == TRUE)
+  if (is_verbose_mode) {
     # Evaluate the message within the calling environment
     eval_message <- glue::glue(msg, .envir = parent.frame())
     cli_inform(eval_message)