+++ tests/testthat/test-DebyeScherrer.R
@@ -1,6 +1,6 @@
 test_that("get Debye-Scherrer width", {
   fn = xrd.getSampleFiles(fileExt='asc')
-  d <- xrd.read.ASC(fn)
+  d <- xrd.import(fn)
   q = xrd.get.PeakStats(d$TwoTheta, d$I, 38.2, Try.Sigma = 0.3)
 
   # expect Au layer to be 14.4nm