+++ tests/testthat/test-findPeak.R
@@ -1,4 +1,4 @@
-filename = xrd.getSampleFiles()[1]
+filename = xrd.getSampleFiles('asc')
 d = xrd.import(filename)
 
 test_that("Find largest peak", {
@@ -19,12 +19,12 @@ test_that("Test Peak Position Fit", {
 test_that("Test Peak Background and Amplitude Fit", {
   q = xrd.get.PeakStats(d$TwoTheta, d$I, 38.2)
   expect_equal(q[2], 422,Try.Sigma = 0.3, tolerance=1e-3) # peak amplitude
-  expect_equal(q[1], 890,Try.Sigma = 0.3, tolerance=1e-3)# peak background
+  expect_equal(q[1], 890,Try.Sigma = 0.3, tolerance=1e-3) # peak background
 })
 
 test_that("No peak at 42 or 72, but warning", {
   filename <- xrd.getSampleFiles("asc")
-  d = xrd.read.ASC(filename)
+  d = xrd.import(filename)
 
   # there is no peak in this data
   expect_warning(xrd.find.Peak(d$TwoTheta, d$I,42, thetaDelta=5),