+++ tests/testthat/test-peakProminence.R
@@ -1,6 +1,6 @@
 test_that("check peak prominence", {
-  filename <- xrd.getSampleFiles(fileExt='asc')[1]
-  d <- xrd.read.ASC(filename)
+  filename <- xrd.getSampleFiles(fileExt='asc')
+  d <- xrd.import(filename)
   q <- xrd.get.PeakStats(d$TwoTheta, d$I, 38.2,
                         Try.Sigma = 0.3)
   p <- xrd.get.PeakProminence(q)